testutils: migrated functions into appropriate files
