+++ testutils/rand_order.go
@@ -1,17 +1,9 @@
 package testutils
 
 import (
-	"fmt"
 	"math/rand"
 	"time"
 
-	"github.com/republicprotocol/republic-go/contract"
-
-	"github.com/republicprotocol/republic-go/cmd/darknode/config"
-	"github.com/republicprotocol/republic-go/crypto"
-	"github.com/republicprotocol/republic-go/identity"
-	"github.com/republicprotocol/republic-go/logger"
-	"github.com/republicprotocol/republic-go/ome"
 	"github.com/republicprotocol/republic-go/order"
 )
 
@@ -103,87 +95,4 @@ func LessRandomCoExp(coExp order.CoExp) order.CoExp {
 		Co:  co,
 		Exp: exp,
 	}
-}
-
-// RandomConfigs will generate n configs and b of them are bootstrap node.
-func RandomConfigs(n int, b int) ([]config.Config, error) {
-	configs := []config.Config{}
-
-	for i := 0; i < n; i++ {
-		keystore, err := crypto.RandomKeystore()
-		if err != nil {
-			return configs, err
-		}
-
-		addr := identity.Address(keystore.Address())
-		configs = append(configs, config.Config{
-			Keystore:                keystore,
-			Host:                    "0.0.0.0",
-			Port:                    fmt.Sprintf("%d", 18514+i),
-			Address:                 addr,
-			BootstrapMultiAddresses: identity.MultiAddresses{},
-			Logs: logger.Options{
-				Plugins: []logger.PluginOptions{
-					{
-						File: &logger.FilePluginOptions{
-							Path: fmt.Sprintf("%v.out", addr),
-						},
-					},
-				},
-			},
-			Ethereum: contract.Config{
-				Network: contract.NetworkLocal,
-				URI:     "http://localhost:8545",
-			},
-		})
-	}
-
-	for i := 0; i < n; i++ {
-		for j := 0; j < b; j++ {
-			if i == j {
-				continue
-			}
-			bootstrapMultiAddr, err := identity.NewMultiAddressFromString(fmt.Sprintf("/ip4/%v/tcp/%v/republic/%v", configs[j].Host, configs[j].Port, configs[j].Address))
-			if err != nil {
-				return configs, err
-			}
-			configs[i].BootstrapMultiAddresses = append(configs[i].BootstrapMultiAddresses, bootstrapMultiAddr)
-		}
-	}
-
-	return configs, nil
-}
-
-// Random32Bytes creates a random [32]byte.
-func Random32Bytes() [32]byte {
-	var res [32]byte
-	i := fmt.Sprintf("%d", rand.Int())
-	hash := crypto.Keccak256([]byte(i))
-	copy(res[:], hash)
-
-	return res
-}
-
-// RandomNetworkID generates a random [32]byte array
-func RandomNetworkID() [32]byte {
-	return Random32Bytes()
-}
-
-// RandomComputation generates a random computation with empty epoch hash.
-func RandomComputation() (ome.Computation, error) {
-	buy, sell := RandomBuyOrder(), RandomSellOrder()
-	buyFragments, err := buy.Split(24, 16)
-	if err != nil {
-		return ome.Computation{}, err
-	}
-	sellFragments, err := sell.Split(24, 16)
-	if err != nil {
-		return ome.Computation{}, err
-	}
-	comp := ome.Computation{
-		Buy:  buyFragments[0],
-		Sell: sellFragments[0],
-	}
-	copy(comp.ID[:], crypto.Keccak256(buy.ID[:], sell.ID[:]))
-	return comp, nil
 }
